Use uint8_t as mouse coordinate type in input parsing.
Since the buffer is char, which is signed in most cases (to be precise the
sign is not specified by the C standard), it leads to bad "overflow" like
behaviour and reports negative mouse coordinates. Using uint8_t solves the
problem. Keep in mind that mouse coordinates are still limited by the single
byte minus 33 positions. Which is [0; 222]. That's a limitation of X10 mouse
input protocol we're using.

Fixes #54.